Move duplicated terminology code to a shared project between STU3 and Conformance
