+++ src/Hl7.Fhir.STU3/Hl7.Fhir.STU3.csproj
@@ -3,6 +3,7 @@
 	<Import Project="..\firely-net-sdk.props" />
 	<Import Project="..\platforms.props" />
 	<Import Project="..\Hl7.Fhir.Shims.STU3AndUp\Hl7.Fhir.Shims.STU3AndUp.projitems" Label="Shared" />
+	<Import Project="..\Hl7.Fhir.Shims.Base\Hl7.Fhir.Shims.Base.projitems" Label="Shared" />
 
 	<PropertyGroup>
 		<PackageId>Hl7.Fhir.STU3</PackageId>
@@ -10,6 +11,10 @@
 		<Title>Firely's SDK for working with HL7 FHIR STU3 (3.0.2)</Title>
 		<Description>Firely's SDK for working with HL7 FHIR STU3. This is the root package for the SDK includes core functionality to working with RESTful FHIR servers, POCO classes for FHIR, parsing/serialization of FHIR data and working with conformance data and terminologies.</Description>
 		<PackageTags>HL7;FHIR;Firely;SDK;POCO;Fhir client;Terminology;StructureDefinition;parsing;serialization;FHIR server</PackageTags>
+	</PropertyGroup>
+
+	<PropertyGroup>
+		<DefineConstants>STU3</DefineConstants>
 	</PropertyGroup>
 
 	<ItemGroup>
