--- src/Hl7.Fhir.STU3/Specification/Terminology/TerminologyServiceException.cs
@@ -1,41 +1,0 @@
-﻿/* 
- * Copyright (c) 2016, Firely (info@fire.ly) and contributors
- * See the file CONTRIBUTORS for details.
- * 
- * This file is licensed under the BSD 3-Clause license
- * available at https://raw.githubusercontent.com/FirelyTeam/firely-net-sdk/master/LICENSE
- */
-
-using System;
-
-namespace Hl7.Fhir.Specification.Terminology
-{
-    public class TerminologyServiceException : Exception
-    {
-        public TerminologyServiceException(string message) : base(message)
-        {
-        }
-    }
-
-    public class ValueSetExpansionTooBigException : TerminologyServiceException
-    {
-        //422 - too costly
-        public ValueSetExpansionTooBigException(string message) : base(message)
-        {
-        }
-    }
-
-    public class ValueSetExpansionTooComplexException : TerminologyServiceException
-    {
-        public ValueSetExpansionTooComplexException(string message) : base(message)
-        {
-        }
-    }
-
-    public class ValueSetUnknownException : TerminologyServiceException
-    {
-        public ValueSetUnknownException(string message) : base(message)
-        {
-        }
-    }
-}
