+++ src/Hl7.Fhir.Shims.Base/Specification/Terminology/LocalTerminologyService.cs
@@ -11,14 +11,12 @@
 using Hl7.Fhir.Model;
 using Hl7.Fhir.Rest;
 using Hl7.Fhir.Specification.Source;
-using Hl7.Fhir.Support;
 using Hl7.Fhir.Utility;
 using System;
 using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 using T = System.Threading.Tasks;
 
 namespace Hl7.Fhir.Specification.Terminology
@@ -30,7 +28,6 @@ namespace Hl7.Fhir.Specification.Terminology
         private readonly IAsyncResourceResolver _resolver;
         private readonly ValueSetExpander _expander;
 
-
         public LocalTerminologyService(IAsyncResourceResolver resolver, ValueSetExpanderSettings? expanderSettings = null)
         {
             _resolver = resolver ?? throw Error.ArgumentNull(nameof(resolver));
@@ -40,7 +37,6 @@ namespace Hl7.Fhir.Specification.Terminology
 
             _expander = new ValueSetExpander(settings);
         }
-
 
         /// <summary>
         /// Creates a MultiTerminologyService, which combines a LocalTerminologyService to retrieve the core FHIR resources with custom services to validate some implicit core ValueSets.
@@ -57,7 +53,11 @@ namespace Hl7.Fhir.Specification.Terminology
         {
             var valueset = await _resolver.FindValueSetAsync(canonical).ConfigureAwait(false);
 
+#if STU3
             if (valueset == null && _resolver is IConformanceSource source)
+#else
+            if (valueset == null && _resolver is ICommonConformanceSource source)
+#endif
             {
                 var cs = source.FindCodeSystemByValueSet(canonical);
                 if (cs != null)
@@ -86,7 +86,6 @@ namespace Hl7.Fhir.Specification.Terminology
         ///<inheritdoc />
         public async T.Task<Parameters> ValueSetValidateCode(Parameters parameters, string? id = null, bool useGet = false)
         {
-
             parameters.CheckForValidityOfValidateCodeParams();
 
             var validCodeParams = new ValidateCodeParameters(parameters);
@@ -124,7 +123,7 @@ namespace Hl7.Fhir.Specification.Terminology
             }
 #pragma warning disable CS0618 // Only catched here to not change to public interface of ValueSetExpander
             catch (TerminologyServiceException e)
-#pragma warning restore CS0618 
+#pragma warning restore CS0618
             {
                 throw new FhirOperationException(e.Message, (HttpStatusCode)422);
             }
@@ -284,6 +283,7 @@ namespace Hl7.Fhir.Specification.Terminology
                 result.Add("message", new FhirString(messages.ToString().TrimEnd()));
             return result;
         }
+
         private async T.Task messageForCodeNotFound(ValueSet vs, string? system, string codeLabel, StringBuilder messages)
         {
             if (system is not null && await isValueSet(system).ConfigureAwait(false))
@@ -295,7 +295,7 @@ namespace Hl7.Fhir.Specification.Terminology
                 messages.AppendLine($"{codeLabel} does not exist in the value set '{vs.Title ?? vs.Name}' ({vs.Url})");
             }
 
-            async Task<bool> isValueSet(string system)
+            async T.Task<bool> isValueSet(string system)
             {
                 // First, conduct a quick initial check, and if that fails, proceed with a more comprehensive approach.
                 return (system.Contains(@"/ValueSet/") || await _resolver.FindValueSetAsync(system).ConfigureAwait(false) is not null);
