+++ src/Hl7.Fhir.Shims.Base/Specification/Terminology/TerminologyServiceException.cs
